Split testing/mocking into testing/core/mocking and testing/react.
