+++ src/testing/core/index.ts
@@ -3,12 +3,12 @@ export {
   mockSingleLink,
   MockedResponse,
   ResultFunction
-} from '../mocking/mockLink';
+} from './mocking/mockLink';
 export {
   MockSubscriptionLink,
   mockObservableLink
-} from '../mocking/mockSubscriptionLink';
-export { createMockClient } from '../mocking/mockClient';
-export { default as subscribeAndCount } from '../subscribeAndCount';
-export { itAsync } from '../itAsync';
-export * from '../withConsoleSpy';
+} from './mocking/mockSubscriptionLink';
+export { createMockClient } from './mocking/mockClient';
+export { default as subscribeAndCount } from './subscribeAndCount';
+export { itAsync } from './itAsync';
+export * from './withConsoleSpy';
