+++ src/testing/index.ts
@@ -1,3 +1,3 @@
 import '../utilities/globals';
-export { MockedProvider, MockedProviderProps } from './mocking/MockedProvider';
+export { MockedProvider, MockedProviderProps } from './react/MockedProvider';
 export * from './core';
