+++ src/testing/react/__tests__/MockedProvider.test.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
+import { DocumentNode } from 'graphql';
 import { render, wait } from '@testing-library/react';
 import gql from 'graphql-tag';
 
-import { itAsync } from '../../itAsync';
+import { itAsync, MockedResponse, MockLink } from '../../core';
 import { MockedProvider } from '../MockedProvider';
-import { MockedResponse, MockLink } from '../mockLink';
-import { DocumentNode } from 'graphql';
 import { useQuery } from '../../../react/hooks';
 import { InMemoryCache } from '../../../cache';
 import { ApolloLink } from '../../../link/core';
